chore: add new font and moved query data
